+++ src/coordinate-canvas/core/config.py
