+++ src/coordinate-canvas/core/line_builder.py
